+++ packages/contracts/src/interfaces/IFluidLocker.sol
@@ -48,7 +48,7 @@ interface IFluidLocker {
      * @param nonce nonce associated to the signature provided by Stack
      * @param stackSignature stack signature containing necessary info to update units
      */
-    function claim(uint256 programId, uint128 totalProgramUnits, uint256 nonce, bytes memory stackSignature) external;
+    function claim(uint256 programId, uint256 totalProgramUnits, uint256 nonce, bytes memory stackSignature) external;
 
     /**
      * @notice Batch update this locker units within the given programs identifier's GDA pools
@@ -59,7 +59,7 @@ interface IFluidLocker {
      */
     function claim(
         uint256[] memory programIds,
-        uint128[] memory totalProgramUnits,
+        uint256[] memory totalProgramUnits,
         uint256[] memory nonces,
         bytes[] memory stackSignatures
     ) external;
